refactor(me): extract image byte sniffing into shared util
Move the raster signature checks out of metadata.ts into a pure helper
so they can be reused outside the og:image validator, and add ICO/SVG
detection for the upcoming link-card favicon support.
